ok consultas no servidor, antes de separar layers
