+++ app/src/main/java/br/com/israelermel/topranked/di/modules/repositories/RepositoriesModule.kt
@@ -1,14 +1,14 @@
 package br.com.israelermel.topranked.di.modules.repositories
 
-import br.com.israelermel.data.db.RepoDatabase
-import br.com.israelermel.data.repository.RemoteGitHubRepositoriesImpl
-import br.com.israelermel.domain.repository.IGitHubRepositoriesRepository
+import br.com.israelermel.database.configs.GitHubDatabase
+import br.com.israelermel.database.configs.RemoteReposImpl
+import br.com.israelermel.domain.repository.IReposRepository
 import org.koin.dsl.bind
 import org.koin.dsl.module
 
 val repositoriesModule = module {
 
-    single { RepoDatabase.getInstance(get()) }
-    single { RemoteGitHubRepositoriesImpl(get(), get()) } bind IGitHubRepositoriesRepository::class
+    single { GitHubDatabase.getInstance(get()) }
+    single { RemoteReposImpl(get(), get()) } bind IReposRepository::class
 
 }
