+++ app/src/main/java/br/com/israelermel/topranked/di/modules/usecases/UseCaseModule.kt
@@ -1,11 +1,11 @@
 package br.com.israelermel.topranked.di.modules.usecases
 
-import br.com.israelermel.domain.usecase.repositories.GetGithubRepositoriesUseCase
+import br.com.israelermel.domain.usecase.repositories.GetReposLanguageKotlinUseCase
 import org.koin.dsl.module
 
 val useCaseModule = module {
 
     // GitHubRepositories
-    factory { GetGithubRepositoriesUseCase(get()) }
+    factory { GetReposLanguageKotlinUseCase(get()) }
 
 }
