--- data/src/main/java/br/com/israelermel/data/db/RemoteKeys.kt
@@ -1,11 +1,0 @@
-package br.com.israelermel.data.db
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "remote_keys")
-data class RemoteKeys(
-    @PrimaryKey val repoId: Long,
-    val prevKey: Int?,
-    val nextKey: Int?
-)
