--- data/src/main/java/br/com/israelermel/data/db/RepoDatabase.kt
@@ -1,39 +1,0 @@
-package br.com.israelermel.data.db
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import br.com.israelermel.domain.models.repositories.ReposEntity
-
-@Database(
-    entities = [ReposEntity::class, RemoteKeys::class],
-    version = 3,
-    exportSchema = false
-)
-abstract class RepoDatabase : RoomDatabase() {
-
-    abstract fun reposDao(): RepoDao
-    abstract fun remoteKeysDao(): RemoteKeysDao
-
-    companion object {
-
-        @Volatile
-        private var INSTANCE: RepoDatabase? = null
-
-        fun getInstance(context: Context): RepoDatabase =
-            INSTANCE ?: synchronized(this) {
-                INSTANCE
-                    ?: buildDatabase(context).also { INSTANCE = it }
-            }
-
-        private fun buildDatabase(context: Context) =
-            Room.databaseBuilder(
-                context.applicationContext,
-                RepoDatabase::class.java, "Github.db"
-            ).fallbackToDestructiveMigration()
-                .build()
-    }
-
-
-}
