--- data/src/main/java/br/com/israelermel/data/repository/RemoteGitHubRepositoriesImpl.kt
@@ -1,44 +1,0 @@
-package br.com.israelermel.data.repository
-
-import androidx.paging.ExperimentalPagingApi
-import androidx.paging.Pager
-import androidx.paging.PagingConfig
-import androidx.paging.PagingData
-import br.com.israelermel.data.db.RepoDatabase
-import br.com.israelermel.data.networking.api.RepositoriesApi
-import br.com.israelermel.data.paging.GithubRemoteMediator
-import br.com.israelermel.domain.models.repositories.GitHubRepositoriesRequest
-import br.com.israelermel.domain.models.repositories.ReposEntity
-import br.com.israelermel.domain.repository.IGitHubRepositoriesRepository
-import kotlinx.coroutines.flow.Flow
-
-
-class RemoteGitHubRepositoriesImpl(
-    private val repositoriesApi: RepositoriesApi,
-    private val database: RepoDatabase
-) : IGitHubRepositoriesRepository {
-
-
-    @OptIn(ExperimentalPagingApi::class)
-    override suspend fun getReposLanguageKotlinResultStream(request: GitHubRepositoriesRequest): Flow<PagingData<ReposEntity>> {
-
-        val queryFilterByKotlin = "kotlin"
-        val pagingSourceFactory = { database.reposDao().reposByLanguage(queryFilterByKotlin) }
-
-        return Pager(
-            config = PagingConfig(
-                pageSize = GithubRemoteMediator.NETWORK_PAGE_SIZE,
-                enablePlaceholders = false,
-                initialLoadSize = GithubRemoteMediator.NETWORK_PAGE_SIZE + 10
-            ),
-            remoteMediator = GithubRemoteMediator(
-                query = queryFilterByKotlin,
-                service = repositoriesApi,
-                repoDatabase = database
-            ),
-            pagingSourceFactory = pagingSourceFactory
-        ).flow
-    }
-
-
-}
