--- domain/src/main/java/br/com/israelermel/domain/models/repositories/RepositoriesBo.kt
@@ -1,11 +1,0 @@
-package br.com.israelermel.domain.models.repositories
-
-data class RepositoriesBo(
-    val id: Long,
-    val fullName: String? = "",
-    val language: String? = "",
-    val forksCount: Int? = 0,
-    val stargazersCount: Int? = 0,
-    val login: String = "",
-    val avatarUrl: String = ""
-)
