+++ domain/src/main/java/br/com/israelermel/domain/repository/IReposRepository.kt
@@ -1,0 +1,9 @@
+package br.com.israelermel.domain.repository
+
+import androidx.paging.PagingData
+import br.com.israelermel.domain.models.repositories.ReposEntity
+import kotlinx.coroutines.flow.Flow
+
+interface IReposRepository {
+    suspend fun getReposLanguageKotlinResultStream(): Flow<PagingData<ReposEntity>>
+}
