+++ features/feature-top-ranked/build.gradle
@@ -21,4 +21,5 @@ dependencies {
     domainModule()
     testingCoreUnitTestModule()
     librariesArchThemes()
+    librariesArchExtensions()
 }
