+++ features/feature-top-ranked/src/main/java/br/com/israelermel/feature_top_ranked/scenes/TopRankedKotlinRepositoriesViewModel.kt
@@ -4,38 +4,38 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import androidx.paging.cachedIn
 import br.com.israelermel.domain.exceptions.RepositoriesException
-import br.com.israelermel.domain.models.repositories.GitHubRepositoriesRequest
 import br.com.israelermel.domain.states.RequestResult
-import br.com.israelermel.domain.usecase.repositories.GetGithubRepositoriesUseCase
-import br.com.israelermel.feature_top_ranked.states.GitHubRepositoriesState
+import br.com.israelermel.domain.usecase.repositories.GetReposLanguageKotlinUseCase
+import br.com.israelermel.feature_top_ranked.states.ReposResultState
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 
 class TopRankedKotlinRepositoriesViewModel(
-    private val getGithubRepositoriesUseCase: GetGithubRepositoriesUseCase
+    private val getReposLanguageKotlinUseCase: GetReposLanguageKotlinUseCase
 ) : ViewModel() {
 
-    private val _resultState = MutableLiveData<GitHubRepositoriesState>()
+    private val _resultState = MutableLiveData<ReposResultState>()
 
-    val resultState: LiveData<GitHubRepositoriesState>
+    val resultResultState: LiveData<ReposResultState>
         get() = _resultState
 
 
-    fun getGitHubRepositories(resquest: GitHubRepositoriesRequest) {
+    fun getGitHubRepositories() {
         viewModelScope.launch {
 
-            when (val repositories = getGithubRepositoriesUseCase.execute(resquest)) {
+            when (val repositories = getReposLanguageKotlinUseCase.execute()) {
                 is RequestResult.Success -> {
 
-                    repositories.result.collectLatest {
-                        _resultState.postValue(GitHubRepositoriesState.Success(it))
+                    repositories.result.cachedIn(viewModelScope).collectLatest {
+                        _resultState.postValue(ReposResultState.Success(it))
                     }
                 }
 
                 is RequestResult.Failure -> {
                     _resultState.postValue(
-                        GitHubRepositoriesState.Error(
+                        ReposResultState.Error(
                             error = repositories.throwable as RepositoriesException
                         )
                     )
