--- features/feature-top-ranked/src/main/java/br/com/israelermel/feature_top_ranked/states/GitHubRepositoriesState.kt
@@ -1,12 +1,0 @@
-package br.com.israelermel.feature_top_ranked.states
-
-import androidx.paging.PagingData
-import br.com.israelermel.domain.exceptions.RepositoriesException
-import br.com.israelermel.domain.models.repositories.ReposEntity
-
-sealed class GitHubRepositoriesState {
-    object Loading : GitHubRepositoriesState()
-    data class Success(val repositories: PagingData<ReposEntity>) : GitHubRepositoriesState()
-    data class Error(val error: RepositoriesException) : GitHubRepositoriesState()
-
-}
