+++ features/feature-top-ranked/src/main/java/br/com/israelermel/feature_top_ranked/states/TopRankedUserEvent.kt
@@ -1,7 +1,5 @@
 package br.com.israelermel.feature_top_ranked.states
 
-import br.com.israelermel.domain.models.repositories.GitHubRepositoriesRequest
-
 sealed class TopRankedUserEvent {
-    data class GetTopRankedUserEvent(val getTopRankedProjects: GitHubRepositoriesRequest) : TopRankedUserEvent()
+    object GetTopRankedUserEvent : TopRankedUserEvent()
 }
