+++ infra/infra-database/build.gradle
@@ -1,0 +1,22 @@
+plugins {
+    id 'com.android.library'
+    id 'kotlin-android'
+    id 'kotlin-android-extensions'
+    id 'kotlin-kapt'
+}
+
+apply from: "$rootProject.rootDir/groups.gradle"
+
+dependencies {
+    stdKotlin()
+    kotlinKtx()
+    support()
+    coroutines()
+    paging()
+    room()
+    networking()
+
+    //Modules
+    domainModule()
+    dataModule()
+}
