+++ infra/infra-database/src/main/java/br/com/israelermel/database/remotekeys/RemoteKeys.kt
@@ -1,0 +1,11 @@
+package br.com.israelermel.database.remotekeys
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "remote_keys")
+data class RemoteKeys(
+    @PrimaryKey val repoId: Long,
+    val prevKey: Int?,
+    val nextKey: Int?
+)
