+++ libraries/library-arch-extensions/src/test/java/br/com/israelermel/library_arch_extensions/ExampleUnitTest.kt
@@ -1,0 +1,17 @@
+package br.com.israelermel.library_arch_extensions
+
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
